HARMONY-1160: Interim commit removing cli stuff and query without scrolling
